# le-backend/docs/openapi_enhanced.yaml
openapi: 3.0.3
info:
  title: LC Workflow API - Enhanced
  description: |
    Enhanced LC Workflow API with comprehensive system stability improvements, 
    security features, monitoring capabilities, and data synchronization.
    
    ## Features
    
    - **Enhanced File Management**: Secure file upload with malware scanning and encryption
    - **Folder Organization**: Automatic document organization by type
    - **Health Monitoring**: Comprehensive system health checks and metrics
    - **Data Synchronization**: Real-time updates and cache management
    - **Security**: Advanced authentication, authorization, and audit logging
    - **Feature Flags**: Gradual rollout and A/B testing capabilities
    
    ## Authentication
    
    All endpoints require JWT authentication unless specified otherwise.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Error Handling
    
    The API uses standardized error responses with correlation IDs for tracking.
    All errors include suggested actions and detailed information.
    
    ## Rate Limiting
    
    - File Upload: 100 requests/hour per user
    - General API: 1000 requests/hour per user
    - Health Checks: 60 requests/minute (no user limit)
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /folders/:
    get:
      tags:
        - Folders
      summary: Get folders
      description: Retrieve folders with filtering and pagination
      operationId: getFolders
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by parent folder
        - name: application_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by application
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Page number for pagination
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: Folders retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/FolderResponse'
                  - $ref: '#/components/schemas/PaginatedFolderResponse'

    post:
      tags:
        - Folders
      summary: Create folder
      description: Create a new folder
      operationId: createFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FolderCreateRequest'
      responses:
        '201':
          description: Folder created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
...
  /folders/application/{application_id}/hierarchy:
    get:
      tags:
        - Folders
      summary: Get application folder hierarchy
      description: Get complete folder structure for an application
      operationId: getApplicationFolderHierarchy
      parameters:
        - name: application_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Folder hierarchy retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderHierarchyResponse'

  # Health Monitoring Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        original_filename:
          type: string
        display_name:
          type: string
        file_path:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        uploaded_by:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
          nullable: true
        folder_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        download_url:
          type: string
          format: uri
          description: Secure download URL (when applicable)

    # Folder Management Schemas
    FolderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        application_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileResponse'
    FolderCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        application_id:
          type: string
          format: uuid
          nullable: true
    PaginatedFolderResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FolderResponse'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
    FolderHierarchyResponse:
      type: object
      properties:
        application_id:
          type: string
          format: uuid
        root_folder:
          $ref: '#/components/schemas/FolderNode'
        total_folders:
          type: integer
        total_files:
          type: integer
    FolderNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        files_count:
          type: integer
        children:
          type: array
          items:
            $ref: '#/components/schemas/FolderNode'

    # Health Monitoring Schemas
...